cleanup of classes being located in other classes
